requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
compress-fasttext==0.1.5
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
gensim==4.4.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pillow==12.0.0
protobuf==4.25.8
psutil==7.1.3
PyMuPDF==1.26.7
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentencepiece==0.2.1
setuptools==70.2.0
six==1.17.0
smart_open==7.5.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1+cpu
torchaudio==2.9.1+cpu
torchvision==0.24.1+cpu
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
wrapt==2.0.1
